// acs-web/AcsStatsWeb/ClientApp/src/app/modules/fielding-records/models/fielding-overall.model.ts
export interface FieldingCareerRecordDto {
  playerId: number;
  name: string;
  sortNamePart: string;
  team: string;
  debut: string
  end: string
  opponents: string;
  year: string;
  matches: number;
  innings: number;
  ground: string;
  countryName: string;
  dismissals: number;
  wicketKeeperDismissals: number;
  caught: number;
  stumpings: number;
  caughtKeeper: number;
  caughtFielder: number;
  bestDismissals: number;
  bestCaughtFielder: number;
  bestCaughtKeeper: number;
  bestStumpings: number;
}